Update tasks and build modules
Remove module loading in preference of having each build module define its
own method for building. Although I like the idea of using the loading
method, there are far too many other build tools for me to be able to
sanely cover all of these in the loading method. Further, the build tools
are not the only set of commands that differ based on which module is
loaded (or at least, won't be...)