Update ``echo`` task: remove print and subprocess
If anyone wants to "echo" to screen, they can (and probably will) just
call ``print``. So the ``echo`` task does not need to do this for them.
However, it is still desired to have some quick facility to concisely
write to a file. Further, the method in which echo was writing to file was
silly and quite un-pythonic. As a result of the removal of subprocess from
echo, the echo test(s) needed to be updated to remove the test for the
``\n`` character