Refactor: Add NOOP
Change all modules to return a tuple (of tuples) of functions that would be
executed.  The resulting tuples look similar to:

    ((references_to_real_function, dictionary_of_arguments), ...)

The tuple can be applied by calling the function reference, passing the
dictionary (expanded with `**` magic). Such a function is defined as
`xnt.tasks.__apply__`.

The following are affected modules:

*   `xnt.tasks`

*   `xnt.build.cc`

*   `xnt.build.make`

*   `xnt.build.tex`

*   `xnt.vcs.git`

    -   Patch includes added tests

*   `xnt.vcs.hg`

    -   Patch includes added tests

*   `xnt.vcs.cvs`

    -   Patch includes added tests