Move parameter passing
Full parameter passing (converting `key=value` to `{'key': 'value'}`)was
happening in the `xnt.core_tasks` module, and attempting to merge the
dictionaries. Move the former functionality to the `xnt.xenant` module and
pass the result of the first pass to the `xnt.xntcall` function, which will
merge the loaded properties with the parsed properties.